Replace manage with new deploy tool and update README to reflect new deployment method
This replaces the previous manage tool and the README which had the user
step through each step calling manage to deploy Birch Girder. It also
removes the need for any manual setup steps for GitHub or AWS.

The new deploy script and the associated README instructions prompt the user
for any information needed and converge GitHub and AWS to the state asserted
in the config.