Switch from GitHub services to hook.io for SNS integration
Now that GitHub [has retired GitHub Services](https://developer.github.com/changes/2018-04-25-github-services-deprecation/)
we need to replace the GitHub Services "amazonsns" integration with something
so that comments in GitHub issues can trigger birch-girder to send an email
to a recipient.

This removes the "amazonsns" GitHub Services integration and replaces it with
a hook.io integration. Now a GitHub issue comment will trigger a GitHub
webhook, which points to a hook.io endpoint. The hook.io code will check
the signature on the request to make sure it came from an allowed GitHub
repository. It will then publish the payload into the configured Amazon SNS
topic so that birch-girder can process it.